Migrate test_circuits.py to new tests
